requirements.txt
JSON-minify==0.3.0
PyMySQL==1.1.1
aiofiles==24.1.0
aiohttp==3.10.11
apscheduler==3.10.4
asgiref==3.8.1
asyncer==0.0.8
chromadb==0.5.15
coverage==7.6.4
cryptography==43.0.3
eval-type-backport==0.2.0
fastapi==0.115.4
gunicorn==23.0.0
jinja2==3.1.4
jionlp==1.5.17
jsonnet-binary==0.17.0
jsonschema==4.23.0
jieba==0.42.1
httpx==0.27.2
langchain-community==0.3.5
langchain-core==0.3.15
langchain-openai==0.2.5
langchain==0.3.7
limits==3.7.0
minio==7.2.11
numpy==1.26.4
ollama==0.4.4
openai==1.57.0
openpyxl==3.1.5
paramiko==3.4.0
pgvector==0.3.6
psycopg2-binary==2.9.9
pydantic==2.9.2
python-magic==0.4.27
pymongo==4.10.1
python-dotenv==1.0.0
python-jsonpath==1.2.0
python-multipart==0.0.9
pytz==2024.2
pyyaml==6.0.2
rank-bm25==0.2.2
redis==5.2.0
requests==2.32.3
sglang==0.4.0.post1
sortedcontainers==2.4.0
spark-ai-python==0.4.5
sqlalchemy==2.0.35
starlette==0.41.2
tiktoken==0.8.0
untruncate-json==1.0.0
uvicorn==0.21.0
watchdog==5.0.3